+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file, following t
 
 ## [Unreleased]
 
+## [0.7.0] - 2024-03-15
+
 - Changing domains while keeping data (`setDomains`)
 - Adjust zoom behavior when xDomain or yDomain changes size
 - Styling via CSS